using TestHelpers and start additional one
